parsing fonera freifunk variant, verbose option
The fonera freifunk variant uses a redirect on the first page, we also
match this now to determine it is a freifunk device.
The fonera freifunk variant uses a slightly different interface config
that doesn't have a list of interfaces at the start. We now loop over
the interfaces and check if we missed any.
We now check for localnet and link local address in addition to rfc1918.
Add verbose option to spider/parser.